Moved options-related structs and enums to separate module...
